moving GlyphSubstitutionTable into its own package
