+++ fontbox/src/main/java/org/apache/fontbox/ttf/TTFParser.java
@@ -20,6 +20,8 @@ import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
 
+import org.apache.fontbox.ttf.gsub.GlyphSubstitutionTable;
+
 /**
  * TrueType font file parser.
  *
